modified idle state
State machine switches now also whilst the player is pressing the move button to move state. This solved the issue that the player could hold the move button whilst he was performing a skill. But the state machine wont switch to idle until the skill was finished so when it swtiches to idle state after the attack state it can't catcht the move start event in some occassions