fixed bug in character attribute
the character attribute was only calculating its value if it had changes in any of its list. If there are no changes it simply retunrs the calculated value from its last calculation. As it is a scriptable object it was not recalculting the new value if there were changes in the edotor.