fixed bug in attack state
The player was freezin sometimes on animation changes. The bug was that i was checking the m_IsPressingBasicAttack in the change state function. But this if statement is actually handled from the funciton which is calling the change state function. The bool was sometimes changing its value after the change state method was called.